Fix: Invalid track sorting on artist select
When playing an artist from the artist category, it alphabetically
sorts by album + track but did not account for padding the start of
the track. As a result it would sort track 10 immediately after track 1.

Additionally, types for musicmetadata were added and it was discovered
that not all metadata fed into the data store were string types. This
change converts the track from a number to a string, and now
correctly expects the duration to be a number instead of a string.