Fix: Deselect a category
Clicking on the top row of a catagory stylized as "[All - 1022]" is
effectively a deselection of that category. This currently applies to
artist and album. Because playing all artists or all albums is the same
as playing the entire music library, there's no practical need for it to
return any songs to play. It should just clear the browsing playlist and
restore and categories that have been narrowed down.